Revert "ADOP-2566: Set up migratecase event for migration tool (#958)"
This reverts commit fb45491a858dafa9dc71fa4d723ecb009cdbf38b.
