--- src/main/java/uk/gov/hmcts/reform/adoption/adoptioncase/model/access/SystemUpdateOnlyAccess.java
@@ -1,19 +1,0 @@
-package uk.gov.hmcts.reform.adoption.adoptioncase.model.access;
-
-import com.google.common.collect.HashMultimap;
-import com.google.common.collect.SetMultimap;
-import uk.gov.hmcts.ccd.sdk.api.HasAccessControl;
-import uk.gov.hmcts.ccd.sdk.api.HasRole;
-import uk.gov.hmcts.ccd.sdk.api.Permission;
-
-import static uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole.SYSTEM_UPDATE;
-
-public class SystemUpdateOnlyAccess implements HasAccessControl {
-    @Override
-    public SetMultimap<HasRole, Permission> getGrants() {
-        SetMultimap<HasRole, Permission> grants = HashMultimap.create();
-        grants.putAll(SYSTEM_UPDATE, Permissions.CREATE_READ_UPDATE_DELETE);
-
-        return grants;
-    }
-}
