+++ src/test/java/uk/gov/hmcts/reform/adoption/adoptioncase/caseworker/event/CaseworkerCheckAndSendOrdersTest.java
@@ -1,42 +1,51 @@
 package uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.validation.constraints.NotNull;
+import com.google.common.collect.ImmutableSet;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
+import uk.gov.hmcts.ccd.sdk.ResolvedCCDConfig;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.Event;
+import uk.gov.hmcts.ccd.sdk.api.HasRole;
 import uk.gov.hmcts.ccd.sdk.type.DynamicList;
 import uk.gov.hmcts.ccd.sdk.type.DynamicListElement;
 import uk.gov.hmcts.ccd.sdk.type.ListValue;
-import uk.gov.hmcts.reform.adoption.adoptioncase.event.EventTest;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.AdoptionOrderData;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.ManageOrdersData;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.OrderCheckAndSend;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.SelectedOrder;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.ManageOrdersData;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.AdoptionOrderData;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.OrderData;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.OrderStatus;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.SelectedOrder;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
 import uk.gov.hmcts.reform.adoption.document.CaseDataDocumentService;
 import uk.gov.hmcts.reform.adoption.idam.IdamService;
+import uk.gov.hmcts.reform.idam.client.models.User;
+import uk.gov.hmcts.reform.idam.client.models.UserDetails;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.constraints.NotNull;
 
+import java.lang.reflect.InvocationTargetException;
 import java.time.Clock;
-import java.time.Instant;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.Instant;
 import java.time.ZoneId;
+import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
+import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.platform.commons.util.ReflectionUtils.findMethod;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event.CaseworkerCheckAndSendOrders.CASEWORKER_CHECK_AND_SEND_ORDERS;
@@ -46,7 +55,7 @@ import static uk.gov.hmcts.reform.adoption.testutil.TestConstants.TEST_AUTHORIZA
 import static uk.gov.hmcts.reform.adoption.testutil.TestDataHelper.caseData;
 
 @ExtendWith(MockitoExtension.class)
-class CaseworkerCheckAndSendOrdersTest extends EventTest {
+class CaseworkerCheckAndSendOrdersTest {
 
     @Mock
     private HttpServletRequest httpServletRequest;
@@ -126,7 +135,7 @@ class CaseworkerCheckAndSendOrdersTest extends EventTest {
         when(clock.getZone()).thenReturn(zoneId);
         when(httpServletRequest.getHeader(AUTHORIZATION)).thenReturn(TEST_AUTHORIZATION_TOKEN);
 
-        when(idamService.retrieveUser(TEST_AUTHORIZATION_TOKEN)).thenReturn(getJudgeUser());
+        when(idamService.retrieveUser(TEST_AUTHORIZATION_TOKEN)).thenReturn(getCaseworkerUser());
 
         var result = caseworkerCheckAndSendOrders.aboutToSubmit(caseDetails, caseDetails);
         assertThat(result.getData().getManageOrderList().get(0).getValue().getOrderStatus()).isEqualTo(OrderStatus.SERVED);
@@ -156,6 +165,32 @@ class CaseworkerCheckAndSendOrdersTest extends EventTest {
         return manageOrdersData;
     }
 
+    public static ConfigBuilderImpl<CaseData, State, UserRole> createCaseDataConfigBuilder() {
+        return new ConfigBuilderImpl<>(new ResolvedCCDConfig<>(
+            CaseData.class,
+            State.class,
+            UserRole.class,
+            new HashMap<>(),
+            ImmutableSet.copyOf(State.class.getEnumConstants())
+        ));
+    }
+
+    @SuppressWarnings({"unchecked"})
+    public static <T, S, R extends HasRole> Map<String, Event<T, R, S>> getEventsFrom(
+        final ConfigBuilderImpl<T, S, R> configBuilder) {
+
+        return (Map<String, Event<T, R, S>>) findMethod(ConfigBuilderImpl.class, "getEvents")
+            .map(method -> {
+                try {
+                    method.setAccessible(true);
+                    return method.invoke(configBuilder);
+                } catch (IllegalAccessException | InvocationTargetException e) {
+                    throw new AssertionError("Unable to invoke ConfigBuilderImpl.class method getEvents", e);
+                }
+            })
+            .orElseThrow(() -> new AssertionError("Unable to find ConfigBuilderImpl.class method getEvents"));
+    }
+
     private CaseDetails<CaseData, State> getCaseDetails() {
         return CaseDetails.<CaseData, State>builder()
             .data(caseData())
@@ -176,4 +211,16 @@ class CaseworkerCheckAndSendOrdersTest extends EventTest {
         data.setCheckAndSendOrderDropdownList(DynamicList.builder().listItems(listElements)
                                                   .value(element).build());
     }
+
+    private User getCaseworkerUser() {
+        UserDetails userDetails = UserDetails
+            .builder()
+            .forename("testFname")
+            .roles(Arrays.asList(UserRole.DISTRICT_JUDGE.getRole()))
+            .surname("testSname")
+            .build();
+
+        return new User(TEST_AUTHORIZATION_TOKEN, userDetails);
+    }
+
 }
