+++ src/test/java/uk/gov/hmcts/reform/adoption/adoptioncase/caseworker/event/LocalAuthoritySubmitApplicationTest.java
@@ -1,13 +1,17 @@
-package uk.gov.hmcts.reform.adoption.adoptioncase.event;
+package uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event;
 
+import com.google.common.collect.ImmutableSet;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
+import uk.gov.hmcts.ccd.sdk.ResolvedCCDConfig;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.Event;
+import uk.gov.hmcts.ccd.sdk.api.HasRole;
+import uk.gov.hmcts.reform.adoption.adoptioncase.event.LocalAuthoritySubmitApplication;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
@@ -16,7 +20,12 @@ import uk.gov.hmcts.reform.adoption.common.service.SubmissionService;
 import uk.gov.hmcts.reform.adoption.service.task.EventService;
 import uk.gov.hmcts.reform.ccd.client.model.SubmittedCallbackResponse;
 
+import java.lang.reflect.InvocationTargetException;
+import java.util.HashMap;
+import java.util.Map;
+
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.platform.commons.util.ReflectionUtils.findMethod;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.adoption.adoptioncase.event.LocalAuthoritySubmitApplication.LOCAL_AUTHORITY_SUBMIT;
 import static uk.gov.hmcts.reform.adoption.adoptioncase.model.State.LaSubmitted;
@@ -24,7 +33,7 @@ import static uk.gov.hmcts.reform.adoption.adoptioncase.model.State.Submitted;
 import static uk.gov.hmcts.reform.adoption.testutil.TestDataHelper.caseData;
 
 @ExtendWith(MockitoExtension.class)
-public class LocalAuthoritySubmitApplicationTest extends EventTest {
+public class LocalAuthoritySubmitApplicationTest {
 
     @InjectMocks
     private LocalAuthoritySubmitApplication localAuthoritySubmitApplication;
@@ -69,5 +78,31 @@ public class LocalAuthoritySubmitApplicationTest extends EventTest {
             .data(caseData())
             .id(1L)
             .build();
+    }
+
+    public static ConfigBuilderImpl<CaseData, State, UserRole> createCaseDataConfigBuilder() {
+        return new ConfigBuilderImpl<>(new ResolvedCCDConfig<>(
+            CaseData.class,
+            State.class,
+            UserRole.class,
+            new HashMap<>(),
+            ImmutableSet.copyOf(State.class.getEnumConstants())
+        ));
+    }
+
+    @SuppressWarnings({"unchecked"})
+    public static <T, S, R extends HasRole> Map<String, Event<T, R, S>> getEventsFrom(
+        final ConfigBuilderImpl<T, S, R> configBuilder) {
+
+        return (Map<String, Event<T, R, S>>) findMethod(ConfigBuilderImpl.class, "getEvents")
+            .map(method -> {
+                try {
+                    method.setAccessible(true);
+                    return method.invoke(configBuilder);
+                } catch (IllegalAccessException | InvocationTargetException e) {
+                    throw new AssertionError("Unable to invoke ConfigBuilderImpl.class method getEvents", e);
+                }
+            })
+            .orElseThrow(() -> new AssertionError("Unable to find ConfigBuilderImpl.class method getEvents"));
     }
 }
