+++ src/test/java/uk/gov/hmcts/reform/adoption/caseworker/event/CaseworkerAddNoteTest.java
@@ -1,30 +1,39 @@
-package uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event;
+package uk.gov.hmcts.reform.adoption.caseworker.event;
 
-import jakarta.servlet.http.HttpServletRequest;
+import com.google.common.collect.ImmutableSet;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
+import uk.gov.hmcts.ccd.sdk.ResolvedCCDConfig;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.Event;
+import uk.gov.hmcts.ccd.sdk.api.HasRole;
 import uk.gov.hmcts.ccd.sdk.type.ListValue;
-import uk.gov.hmcts.reform.adoption.adoptioncase.event.EventTest;
+import uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event.CaseworkerCaseNote;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseNote;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
 import uk.gov.hmcts.reform.adoption.idam.IdamService;
+import uk.gov.hmcts.reform.idam.client.models.User;
+import uk.gov.hmcts.reform.idam.client.models.UserDetails;
+import jakarta.servlet.http.HttpServletRequest;
 
+import java.lang.reflect.InvocationTargetException;
 import java.time.Clock;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.platform.commons.util.ReflectionUtils.findMethod;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event.CaseworkerCaseNote.CASEWORKER_ADD_CASE_NOTE;
@@ -32,7 +41,7 @@ import static uk.gov.hmcts.reform.adoption.testutil.TestConstants.TEST_AUTHORIZA
 import static uk.gov.hmcts.reform.adoption.testutil.TestDataHelper.caseData;
 
 @ExtendWith(MockitoExtension.class)
-class CaseworkerAddNoteTest extends EventTest {
+class CaseworkerAddNoteTest {
 
     @Mock
     private HttpServletRequest httpServletRequest;
@@ -104,6 +113,16 @@ class CaseworkerAddNoteTest extends EventTest {
         assertThat(result.getData().getCaseNote()).isNotNull();
     }
 
+
+    public static ConfigBuilderImpl<CaseData, State, UserRole> createCaseDataConfigBuilder() {
+        return new ConfigBuilderImpl<>(new ResolvedCCDConfig<>(
+            CaseData.class,
+            State.class,
+            UserRole.class,
+            new HashMap<>(),
+            ImmutableSet.copyOf(State.class.getEnumConstants())));
+    }
+
     private CaseDetails<CaseData, State> getCaseDetails() {
         final var details = new CaseDetails<CaseData, State>();
         final var data = caseData();
@@ -116,4 +135,31 @@ class CaseworkerAddNoteTest extends EventTest {
 
         return details;
     }
+
+    @SuppressWarnings({"unchecked"})
+    public static <T, S, R extends HasRole> Map<String, Event<T, R, S>> getEventsFrom(
+        final ConfigBuilderImpl<T, S, R> configBuilder) {
+
+        return (Map<String, Event<T, R, S>>) findMethod(ConfigBuilderImpl.class, "getEvents")
+            .map(method -> {
+                try {
+                    method.setAccessible(true);
+                    return method.invoke(configBuilder);
+                } catch (IllegalAccessException | InvocationTargetException e) {
+                    throw new AssertionError("Unable to invoke ConfigBuilderImpl.class method getEvents", e);
+                }
+            })
+            .orElseThrow(() -> new AssertionError("Unable to find ConfigBuilderImpl.class method getEvents"));
+    }
+
+    private User getCaseworkerUser() {
+        UserDetails userDetails = UserDetails
+            .builder()
+            .forename("testFname")
+            .surname("testSname")
+            .build();
+
+        return new User(TEST_AUTHORIZATION_TOKEN, userDetails);
+    }
+
 }
