+++ src/test/java/uk/gov/hmcts/reform/adoption/caseworker/event/CaseworkerAmendCaseTest.java
@@ -1,0 +1,77 @@
+package uk.gov.hmcts.reform.adoption.caseworker.event;
+
+import com.google.common.collect.ImmutableSet;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
+import uk.gov.hmcts.ccd.sdk.ResolvedCCDConfig;
+import uk.gov.hmcts.ccd.sdk.api.Event;
+import uk.gov.hmcts.ccd.sdk.api.HasRole;
+import uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event.CaseworkerAmendCase;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
+
+import java.lang.reflect.InvocationTargetException;
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.platform.commons.util.ReflectionUtils.findMethod;
+import static uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event.CaseworkerAmendCase.CASEWORKER_AMEND_CASE;
+
+/**
+ * Unit Test class targeted for Amend Case Details functionality.
+ * CaseworkerAmendCase configuration class will be used as base in all Tests
+ */
+@ExtendWith(MockitoExtension.class)
+public class CaseworkerAmendCaseTest {
+
+    @InjectMocks
+    CaseworkerAmendCase caseworkerAmendCase;
+
+    /**
+     * Test Scenario: Should be able to ADD Configuration to the CCD Config Builder.
+     */
+    @Test
+    void shouldAddConfigurationToConfigBuilder() {
+        final ConfigBuilderImpl<CaseData, State, UserRole> configBuilder = createCaseDataConfigBuilder();
+        caseworkerAmendCase.configure(configBuilder);
+        assertThat(getEventsFrom(configBuilder).values())
+            .extracting(Event::getId)
+            .contains(CASEWORKER_AMEND_CASE);
+    }
+
+    /**
+     * Helper method to create CCD Config builder object to be used for relevant Tests.
+     *
+     * @return - ConfigBuilderImpl
+     */
+    public static ConfigBuilderImpl<CaseData, State, UserRole> createCaseDataConfigBuilder() {
+        return new ConfigBuilderImpl<>(new ResolvedCCDConfig<>(
+            CaseData.class,
+            State.class,
+            UserRole.class,
+            new HashMap<>(),
+            ImmutableSet.copyOf(State.class.getEnumConstants())
+        ));
+    }
+
+    @SuppressWarnings({"unchecked"})
+    public static <T, S, R extends HasRole> Map<String, Event<T, R, S>> getEventsFrom(
+        final ConfigBuilderImpl<T, S, R> configBuilder) {
+
+        return (Map<String, Event<T, R, S>>) findMethod(ConfigBuilderImpl.class, "getEvents")
+            .map(method -> {
+                try {
+                    method.setAccessible(true);
+                    return method.invoke(configBuilder);
+                } catch (IllegalAccessException | InvocationTargetException e) {
+                    throw new AssertionError("Unable to invoke ConfigBuilderImpl.class method getEvents", e);
+                }
+            })
+            .orElseThrow(() -> new AssertionError("Unable to find ConfigBuilderImpl.class method getEvents"));
+    }
+}
